personality ques added and more
Selected a set of personality questions I like and have saved it in theproject itself so i dont forget. It's only 5 4 now, and I'll only b using 1 que 2 geta sense of how I want the ans of ques 2 engage  with the API to return a character. Once I have that finalized, I'll generalize  the method that works to all 5 questions. Once that is said and done, I'll refine the existing 5 ques around the end of the proejct. Then, I'll add 5more ques, refine again, and then I'll post it 2 my portfolio and have my family play with the quiz.