joined listOfQues and optionstraits json file, and more
I joined and clarified the list of questions  and list of traits per option per mcq; the finalized file is: list-of-mcqs-and-traits.json. Now that that is done, I can move onto the next most exciting thing, presenting the list of ques in a UI, and have it grab an answer, i.e. an anime character the user matches to.