code updated to display selected options in  console.log and more...
Ok, So at this point, even though the code is technically untested, I've updated it s.t. it gets me to the root endpoint , i.e. the landing page. From there, the start quiz btn works just fine and takes me to the list of questions. Once I an option from each required-to-ans mcq and press submit, the values are saved in react's context API file in my project called SelectedOptionsContext. It is also displayed in the ctrl+Shift+I chrome developer tools' console tab. So, now then,

NEXT STEPS: are to take the values filled in the context file to the user-matched-tocharacter.js logic file and match the user to an anime character. Once that happens, I will need to update the context file to update its logic s.t. it resents its value to null in case the user whats to have another go at the questionarre. Anyway, that is next. Then finally, I can use the name of the character that the user matched to to search the info of the character in the aniList api using the charactersearch.js logic file. It's pretty awesome!