got submit and reset btn working
Ok so status: selecting options, submitting options, both things work splendidly and when I click submit, the console in chrome does log the list of traits per option, so awesome! I also added a reset btn which we can use 2 clear the global state and local state.

Next up!: I will have the submitted options interact with the charactersearch.js logic and use the options to synethise a set of traits which I'll the use on that js file 2 produce an anime character the user matches to! Awesome! so much 2 look forward 2!