1st phase of LandingPage.js complete
I am enjoying the landingPage.css file that I have created and also the creation of the styles folder in the project.

NEXT STEPS: to glam up the QuizPage.js to the same degree or more as the LandingPage. After that, I can focus on the logic of the matching user to an anime character, and then creating a basic results page which I can then later on also glma up!