updated the state to store matched character
ok, os now i've updated the selectedUptionsContext so that it stores the info of the matched character so that it can be picked up from the results page when i create it and finish it later today.

NEXT STEPS: I'll b updating usermatchedtocharacter so that it stores and then brings up the stored matched character to print in the console.