Anime Quiz functionally complete!
Ok, so the quiz is now functionally complete. the ladnign page is done. There are 5 questions on the quizPage webpage and also when i press submit, it takes me sucessfully to the resultsPage and also ON TOP OF THAT, get the anime data from the anilist api and shows the data in the results Page. NEXT UP: I'll add a restart quiz btn at the bottom of the resultsPage.