beautified QuizPage and ResultsPage
So, the QuizPage and Results page have been sufficiently beautified for now. The logic is also working soundly!

NEXT: I'll create more questions for the mcqs json file and insert it into the the list of questions so that there is  a greater variety possible. Also, I wud do something about keeping the questions more legible. For now, they're fine but not as legible as i wud like them 2 b.