updated project styles and more!
So, I updated the project styles a little more so that they look more fun and engaging.

NEXT STEPS: I wud now, add 5 morequestions to the mcqs json file and then, submit the proejct to my portfolio. I'll come back 2 it at a later date maybe to add some fun soundless video to it and make the app mobile responsive