Project complete for now!
Ok so, I've updated the code s.t. there are 10 mcqs in there now. Also, i've updated user-matching-to-character process by adding a component of randomness to it. Also, I've added beautified the code the best i could. So, now, I'll add it 2 my portfolio but may come back 2 it to make it mobile compatible and also make it more dynamic